requirements.txt
numpy==1.16.0
tqdm==4.29.0
imutils==0.5.2
Pillow==5.4.1
cmake==3.13.2.post1
